Fixed searchmaps map list setup command
It was changed previously because "ls *.bsp" returns error when maps with weird names are present. But being in the "ag" directory and moving the output of "ls maps/*.bsp" to the "allmaps.txt" file, writes all the maps with the "maps/" prefix, which doesn't work as the plugin just takes the whole line removing the ".bsp" and passes it to the agmap command, and "agmap maps/something" doesn't work obviously.