Refactor run requirements
No more hardcoded requirements. They're now taken from func_button and trigger_multiple
entities that have the `requirement_id` keyvalue.

Requirements are sorted by the value of that key. So an entity with requirement_id 1 will have to be fulfilled before one with requirement_id 2. It can start at 0 or any number you wish, it's just the ascending order what matters.

If two or more entities have the same requirement_id, they can be fulfilled in any order, so assuming 4 triggers where 2 of them have the same id, you can complete the
run in any of these 2 orders:

1. Trigger A with requirement_id 1
2. Trigger B with requirement_id 2
3. Trigger C with requirement_id 2
4. Trigger Z with requirement_id 3

or:

1. Trigger A with requirement_id 1
3. Trigger C with requirement_id 2
2. Trigger B with requirement_id 2
4. Trigger Z with requirement_id 3

Note that you can complete C and B in any order, but both of them are still necessary
to complete the next one.